# Improve performances of batch pathfinding and simulation :rocket: 