--- a/editoast/src/views/v2/path/pathfinding.rs
+++ b/editoast/src/views/v2/path/pathfinding.rs
@@ -15,7 +15,6 @@ use serde::Deserialize;
 use tracing::info;
 use utoipa::ToSchema;
 
-use super::CACHE_PATH_EXPIRATION;
 use crate::core::v2::pathfinding::PathfindingRequest;
 use crate::core::v2::pathfinding::PathfindingResult;
 use crate::core::AsCoreRequest;
@@ -27,12 +26,12 @@ use crate::modelsv2::OperationalPointModel;
 use crate::modelsv2::Retrieve;
 use crate::modelsv2::RetrieveBatch;
 use crate::modelsv2::RetrieveBatchUnchecked;
+use crate::modelsv2::RollingStockModel;
 use crate::modelsv2::TrackSectionModel;
 use crate::redis_utils::RedisClient;
 use crate::redis_utils::RedisConnection;
 use crate::views::get_app_version;
 use crate::views::v2::path::PathfindingError;
-use crate::views::v2::train_schedule::TrainScheduleProxy;
 use editoast_models::DbConnection;
 use editoast_models::DbConnectionPoolV2;
 use editoast_schemas::infra::OperationalPoint;
@@ -95,117 +94,202 @@ pub async fn post(
         infra_id: *infra_id,
     })
     .await?;
+
     Ok(Json(
-        pathfinding_blocks(conn, &mut redis_conn, core, &infra, &path_input).await?,
+        pathfinding_blocks(conn, &mut redis_conn, core, &infra, path_input).await?,
     ))
 }
 
+/// Pathfinding computation given a path input
 async fn pathfinding_blocks(
     conn: &mut DbConnection,
     redis_conn: &mut RedisConnection,
     core: Arc<CoreClient>,
     infra: &Infra,
-    path_input: &PathfindingInput,
+    path_input: PathfindingInput,
 ) -> Result<PathfindingResult> {
-    // Compute unique hash of PathInput
-    let hash = path_input_hash(infra.id, &infra.version, path_input);
+    let mut path = pathfinding_blocks_batch(conn, redis_conn, core, infra, &[path_input]).await?;
+    Ok(path.pop().unwrap())
+}
+
+/// Pathfinding batch computation given a list of path inputs
+async fn pathfinding_blocks_batch(
+    conn: &mut DbConnection,
+    redis_conn: &mut RedisConnection,
+    core: Arc<CoreClient>,
+    infra: &Infra,
+    pathfinding_inputs: &[PathfindingInput],
+) -> Result<Vec<PathfindingResult>> {
+    // Compute hashes of all path_inputs
+    let hashes: Vec<_> = pathfinding_inputs
+        .iter()
+        .map(|input| path_input_hash(infra.id, &infra.version, input))
+        .collect();
+
     // Try to retrieve the result from Redis
-    let result: Option<PathfindingResult> =
-        redis_conn.json_get_ex(&hash, CACHE_PATH_EXPIRATION).await?;
-    if let Some(pathfinding) = result {
-        info!("Hit cache");
-        return Ok(pathfinding);
-    }
-    // If miss cache:
+    let mut pathfinding_results: Vec<Option<PathfindingResult>> =
+        redis_conn.json_get_bulk(&hashes).await?;
+
+    // Report number of hit cache
+    let nb_hit = pathfinding_results.iter().flatten().count();
+    info!(
+        nb_hit,
+        nb_miss = pathfinding_inputs.len() - nb_hit,
+        "Hit cache"
+    );
+
+    // Handle miss cache:
     // 1) extract locations from path items
-    let path_items = path_input.clone().path_items;
-    if path_items.len() <= 1 {
-        return Ok(PathfindingResult::NotEnoughPathItems);
+    let mut to_cache = vec![];
+    let mut pathfinding_requests = vec![];
+    let mut pathfinding_requests_index = vec![];
+    for (index, (pathfinding_result, pathfinding_input)) in pathfinding_results
+        .iter_mut()
+        .zip(pathfinding_inputs)
+        .enumerate()
+    {
+        if pathfinding_result.is_some() {
+            continue;
+        }
+
+        match build_pathfinding_request(pathfinding_input, conn, infra).await? {
+            Ok(pathfinding_request) => {
+                pathfinding_requests.push(pathfinding_request);
+                pathfinding_requests_index.push(index);
+            }
+            Err(result) => {
+                *pathfinding_result = Some(result.clone());
+                to_cache.push((&hashes[index], result));
+            }
+        }
     }
-    let result = extract_location_from_path_items(conn, infra.id, &path_items).await?;
-    let track_offsets = match result {
+
+    // 2) Send pathfinding requests to core
+    let mut futures = vec![];
+    for request in &pathfinding_requests {
+        futures.push(Box::pin(request.fetch(core.as_ref())));
+    }
+    let computed_paths: Vec<_> = futures::future::join_all(futures)
+        .await
+        .into_iter()
+        .collect::<Result<_>>()?;
+
+    for (index, computed_path) in computed_paths.into_iter().enumerate() {
+        let path_index = pathfinding_requests_index[index];
+        to_cache.push((&hashes[path_index], computed_path.clone()));
+        pathfinding_results[path_index] = Some(computed_path);
+    }
+
+    // 3) Put in cache
+    redis_conn.json_set_bulk(&to_cache).await?;
+
+    Ok(pathfinding_results.into_iter().flatten().collect())
+}
+
+async fn build_pathfinding_request(
+    pathfinding_input: &PathfindingInput,
+    conn: &mut DbConnection,
+    infra: &Infra,
+) -> Result<std::result::Result<PathfindingRequest, PathfindingResult>> {
+    let path_items = &pathfinding_input.path_items;
+    if path_items.len() <= 1 {
+        return Ok(Err(PathfindingResult::NotEnoughPathItems));
+    }
+    let track_offsets = match extract_location_from_path_items(conn, infra.id, path_items).await? {
         Ok(track_offsets) => track_offsets,
-        Err(e) => return Ok(e.into()),
+        Err(e) => return Ok(Err(e.into())),
     };
 
     // Check if tracks exist
-    if let Err(pathfinding_result) =
-        check_tracks_from_path_items(conn, infra.id, &track_offsets).await?
-    {
-        return Ok(pathfinding_result);
+    if let Err(err) = check_tracks_from_path_items(conn, infra.id, &track_offsets).await? {
+        return Ok(Err(err));
     }
 
-    // 2) Compute path from core
-    let pathfinding_request = PathfindingRequest {
+    // Create the pathfinding request
+    Ok(Ok(PathfindingRequest {
         infra: infra.id,
         expected_version: infra.version.clone(),
         path_items: track_offsets,
-        rolling_stock_loading_gauge: path_input.rolling_stock_loading_gauge,
-        rolling_stock_is_thermal: path_input.rolling_stock_is_thermal,
-        rolling_stock_supported_electrifications: path_input
+        rolling_stock_loading_gauge: pathfinding_input.rolling_stock_loading_gauge,
+        rolling_stock_is_thermal: pathfinding_input.rolling_stock_is_thermal,
+        rolling_stock_supported_electrifications: pathfinding_input
             .rolling_stock_supported_electrifications
             .clone(),
-        rolling_stock_supported_signaling_systems: path_input
+        rolling_stock_supported_signaling_systems: pathfinding_input
             .rolling_stock_supported_signaling_systems
             .clone(),
-    };
-    let pathfinding_result = pathfinding_request.fetch(core.as_ref()).await?;
-
-    // 3) Put in cache
-    redis_conn
-        .json_set_ex(&hash, &pathfinding_result, CACHE_PATH_EXPIRATION)
-        .await?;
-
-    Ok(pathfinding_result)
+    }))
 }
 
-/// Compute a path given a batch of trainschedule and an infrastructure.
-///
-/// ## Important
-///
-/// If this function was called with the same train schedule, the result will be cached.
-/// If you call this function multiple times with the same train schedule but with another infra, then you must provide a fresh `cache`.
+/// Compute a path given a train schedule and an infrastructure.
 pub async fn pathfinding_from_train(
     conn: &mut DbConnection,
     redis: &mut RedisConnection,
     core: Arc<CoreClient>,
     infra: &Infra,
     train_schedule: TrainSchedule,
-    proxy: Arc<TrainScheduleProxy>,
 ) -> Result<PathfindingResult> {
-    if let Some(res) = proxy.get_pathfinding_result(train_schedule.id) {
-        return Ok(res);
-    }
-
-    // Retrieve rolling stock
-    let rolling_stock_name = train_schedule.rolling_stock_name.clone();
-    let Some(rolling_stock) = proxy
-        .get_rolling_stock(rolling_stock_name.clone(), conn)
-        .await?
-    else {
-        return Ok(PathfindingResult::RollingStockNotFound { rolling_stock_name });
-    };
-
-    // Create the path input
-    let path_input = PathfindingInput {
-        rolling_stock_loading_gauge: rolling_stock.loading_gauge,
-        rolling_stock_is_thermal: rolling_stock.has_thermal_curves(),
-        rolling_stock_supported_electrifications: rolling_stock.supported_electrification(),
-        rolling_stock_supported_signaling_systems: rolling_stock.supported_signaling_systems.0,
-        path_items: train_schedule
-            .path
+    let rolling_stocks =
+        RollingStockModel::retrieve(conn, train_schedule.rolling_stock_name.clone())
+            .await?
             .into_iter()
-            .map(|item| item.location)
-            .collect(),
-    };
+            .map(|rs| (rs.name.clone(), rs))
+            .collect();
 
-    match pathfinding_blocks(conn, redis, core, infra, &path_input).await {
-        Ok(res) => {
-            proxy.set_pathfinding_result(train_schedule.id, res.clone());
-            Ok(res)
-        }
-        err => err,
+    Ok(
+        pathfinding_from_train_batch(conn, redis, core, infra, &[train_schedule], &rolling_stocks)
+            .await?
+            .pop()
+            .unwrap(),
+    )
+}
+
+/// Compute a path given a batch of trainschedule and an infrastructure.
+pub async fn pathfinding_from_train_batch(
+    conn: &mut DbConnection,
+    redis: &mut RedisConnection,
+    core: Arc<CoreClient>,
+    infra: &Infra,
+    train_schedules: &[TrainSchedule],
+    rolling_stocks: &HashMap<String, RollingStockModel>,
+) -> Result<Vec<PathfindingResult>> {
+    let mut results = vec![PathfindingResult::NotEnoughPathItems; train_schedules.len()];
+    let mut to_compute = vec![];
+    let mut to_compute_index = vec![];
+    for (index, train_schedule) in train_schedules.iter().enumerate() {
+        // Retrieve rolling stock
+        let rolling_stock_name = &train_schedule.rolling_stock_name;
+        let Some(rolling_stock) = rolling_stocks.get(rolling_stock_name).cloned() else {
+            let rolling_stock_name = rolling_stock_name.clone();
+            results[index] = PathfindingResult::RollingStockNotFound { rolling_stock_name };
+            continue;
+        };
+
+        // Create the path input
+        let path_input = PathfindingInput {
+            rolling_stock_loading_gauge: rolling_stock.loading_gauge,
+            rolling_stock_is_thermal: rolling_stock.has_thermal_curves(),
+            rolling_stock_supported_electrifications: rolling_stock.supported_electrification(),
+            rolling_stock_supported_signaling_systems: rolling_stock.supported_signaling_systems.0,
+            path_items: train_schedule
+                .path
+                .clone()
+                .into_iter()
+                .map(|item| item.location)
+                .collect(),
+        };
+        to_compute.push(path_input);
+        to_compute_index.push(index);
     }
+
+    for (index, res) in pathfinding_blocks_batch(conn, redis, core, infra, &to_compute)
+        .await?
+        .into_iter()
+        .enumerate()
+    {
+        results[to_compute_index[index]] = res;
+    }
+    Ok(results)
 }
 
 /// Generates a unique hash based on the pathfinding entries.