… editoast/src/views/v2/train_schedule/projection.rs
… editoast/src/views/v2/train_schedule/proxy.rs