## What is done

- Massive reduction of Redis call
- Stop handling the expiration time of Redis entries

close #7913

## Benchmark

We are testing two endpoints using batch simulation and pathfinding:
- Simulation summary
- Project path

Tests are performed with everything cached **locally**. I'm guessing that the gain will be much greater in prod with network overhead.

### Before

- Simulation summary: 1120ms
- Project path: 1400ms

![image](https://github.com/OpenRailAssociation/osrd/assets/15349030/18dbcc36-e107-450a-8d7d-16381ae4e049)


### After

- Simulation summary: 800ms
- Project path: 950ms

![image](https://github.com/OpenRailAssociation/osrd/assets/15349030/00d8ce26-132c-4bd6-b403-191e764f956a)